\name{features}
\alias{features}
\title{Build a matrix of features for SINDy}
\description{
  Takes a raw matrix of data and converts into polynomial features
  }

\usage{

}

\arguments{
  \item{x}{Raw data to be converted into features}
  \item{polyorder}{Order of polynomials (including k-th self products)} 
}
\details{
 Expands raw data into a set of polynomial features.
}

\value{Returns a new matrix of data with features from raw data
}

\author{Rick Dale and Harish Bhat}


\examples{

}

\keyword{misc}
