\name{sindyr-package}
\alias{sindyr-package}

\docType{package}

\title{
  Tools for Sparse Identification of Nonlinear Dynamics for R
}

\description{
  This implements the Brunton et al (2016; PNAS) sparse
    identification algorithm for finding ordinary differential
    equations for a measured system from raw data (SINDy). The library
    includes a set of additional tools for working with raw
    data, with an emphasis on cognitive science applications 
    (Dale and Bhat, in press).
}
\details{
\tabular{ll}{
Package: \tab sindyr\cr
Type: \tab Package\cr
Version: \tab 0.002\cr
Date: \tab 2018-07-04\cr
License: \tab GPL >= 2\cr
}


\code{sindy}:

\code{windowed_sindy}: 

\code{features}: 

\code{finite_differences}: 

\code{finite_difference}: 

}

\author{Rick Dale and Harish S. Bhat}

\references{
  
  Dale, R. and Bhat, H. S. (in press). Equations of mind: data science for inferring
  nonlinear dynamics of socio-cognitive systems. Cognitive Systems Research.
  
  Brunton, S. L., Proctor, J. L., and Kutz, J. N. (2016). Discovering governing equations from data by sparse identification of nonlinear dynamical systems. Proceedings of the National Academy of Sciences, 113(15), 3932-3937.  
    
}

\keyword{ package }

\examples{

#txt = "here is a raw raw raw string, literally"
#res = text_rqa(txt,typ = "string")
#plot_rp(res$RP)

}
